+++ medaugmentx/__init__.py
@@ -10,7 +10,7 @@ from medaugmentx.core import (
     Transform,
 )
 
-__version__ = "0.4.0"
+__version__ = "0.5.0"
 
 __all__ = [
     "__version__",